requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==5.4.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
PyJWT==2.0.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
whitenoise==5.2.0
